third_party/WebKit/LayoutTests/fast/media/media-query-overflow-value.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description("Test clamping of overflowing media query lengths")

shouldBeFalse("matchMedia('(min-width: 9999999999px)').matches");
shouldBeTrue("matchMedia('(max-width: 9999999999px)').matches");
shouldBeFalse("matchMedia('(min-device-width: 9999999999px)').matches");
shouldBeTrue("matchMedia('(max-device-width: 9999999999px)').matches");
shouldBeFalse("matchMedia('(min-height: 9999999999px)').matches");
shouldBeTrue("matchMedia('(max-height: 9999999999px)').matches");
shouldBeFalse("matchMedia('(min-device-height: 9999999999px)').matches");
shouldBeTrue("matchMedia('(max-device-height: 9999999999px)').matches");
</script>
