third_party/blink/web_tests/platform/generic/fast/media/media-query-overflow-value-expected.txt

Test clamping of overflowing media query lengths

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS matchMedia('(min-width: 9999999999px)').matches is false
PASS matchMedia('(max-width: 9999999999px)').matches is true
PASS matchMedia('(min-device-width: 9999999999px)').matches is false
PASS matchMedia('(max-device-width: 9999999999px)').matches is true
PASS matchMedia('(min-height: 9999999999px)').matches is false
PASS matchMedia('(max-height: 9999999999px)').matches is true
PASS matchMedia('(min-device-height: 9999999999px)').matches is false
PASS matchMedia('(max-device-height: 9999999999px)').matches is true
PASS successfullyParsed is true

TEST COMPLETE
